requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
dnspython==2.7.0
pydantic==2.10.0
beautifulsoup4==4.12.3
certifi>=2024.0.0
